feat: add a request to api for real data
