+++ src/components/IngredientList/IngredientList.jsx
@@ -1,12 +1,14 @@
 import React from 'react';
 import IngredientCard from '../IngredientCard/IngredientCard';
 import styles from './IngredientList.module.css';
-import { data } from '../../utils/data';
+import {FoodDataContext} from '../../contexts/foodDataContext';
 
 export default function IngredientList(props) {
+  const foodData = React.useContext(FoodDataContext);
+
   return(
      <ul className={`${styles.list} mb-10`}>
-       {data.map(item => {
+       {foodData.map(item => {
          if (item.type === props.type) {
            return (
              <IngredientCard
