+++ src/contexts/foodDataContext.js
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const FoodDataContext = React.createContext();
